add post search hook to search results, re #4771
