+++ arches/app/search/components/base.py
@@ -37,7 +37,7 @@ class BaseSearchFilter():
 
         pass
 
-    def post_search_hook(self, search_results_object, results):
+    def post_search_hook(self, search_results_object, results, permitted_nodegroups):
         """
         code to run after the search results have been retrieved
